Completed the Full-Stack Internship assignment
